without summary and saver

+++ lib/preprocess.py
@@ -15,7 +15,7 @@ def processLeft(img):
     crop = cfg.param.crop_shape
     img = img[0:crop[0],0:crop[1]]
     img = (img-np.mean(img))/np.std(img)
-    return img
+    return img.astype(np.float32)
 
 def processRight(img):
     return processLeft(img)
@@ -25,7 +25,7 @@ def processDisp(img):
     '''
     crop = cfg.param.crop_shape
     img = img[0:crop[0],0:crop[1]]
-    return img
+    return img.astype(np.float32)
